requirements.txt
streamlit
pandas
geopandas
shapely.geometry
pycountry
reverse_geocoder